requirements.txt
Django==5.0.11
dj-database-url==2.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
wheel==0.38.1
whitenoise==6.6.0
python-dotenv==1.0.1
python-decouple==3.8